+++ lib/src/base/scroll_bar.dart
@@ -1,6 +1,35 @@
 import 'package:arna/arna.dart';
 
+/// An Arna-styled scrollbar.
+///
+/// To add a scrollbar to a [ScrollView], wrap the scroll view
+/// widget in a [ArnaScrollbar] widget.
+///
+/// {@macro flutter.widgets.Scrollbar}
+///
+///
+/// {@tool dartpad}
+/// When isAlwaysShown is true, the scrollbar thumb will remain visible.
+/// This requires that a ScrollController is provided to controller, or that
+/// the PrimaryScrollController is available.
+///
+/// {@end-tool}
+///
+/// See also:
+///
+///  * [RawScrollbar], a basic scrollbar that fades in and out, extended
+///    by this class to add more animations and behaviors.
+///  * [ListView], which displays a linear, scrollable list of children.
+///  * [GridView], which displays a 2 dimensional, scrollable array of children.
 class ArnaScrollbar extends RawScrollbar {
+  /// Creates an Arna-styled scrollbar that by default will connect to the
+  /// closest Scrollable descendant of [child].
+  ///
+  /// The [child] should be a source of [ScrollNotification] notifications,
+  /// typically a [Scrollable] widget.
+  ///
+  /// If the [controller] is null, the default behavior is to
+  /// enable scrollbar dragging using the [PrimaryScrollController].
   const ArnaScrollbar({
     Key? key,
     required Widget child,