+++ lib/src/base/side_scaffold.dart
@@ -1,22 +1,38 @@
 import 'package:arna/arna.dart';
 
+/// Implements the side view layout structure.
+/// See also:
+///
+///  * [ArnaHeaderBar], which is a horizontal bar shown at the top of the app.
 class ArnaSideScaffold extends StatefulWidget {
-  final Widget headerBarLeading;
-  final String? title;
-  final Widget headerBarTrailing;
-  final List<NavigationItem> items;
-  final ValueChanged<int>? onItemSelected;
-  final int currentIndex;
-
+  /// Creates a side view structure in the Arna style.
   const ArnaSideScaffold({
     Key? key,
-    this.headerBarLeading = const SizedBox.shrink(),
+    this.headerBarLeading,
     this.title,
-    this.headerBarTrailing = const SizedBox.shrink(),
+    this.headerBarTrailing,
     required this.items,
     this.onItemSelected,
     this.currentIndex = 0,
   }) : super(key: key);
+
+  /// The leading widget laid out within the header bar.
+  final Widget? headerBarLeading;
+
+  /// The title displayed in the header bar.
+  final String? title;
+
+  /// The trailing widget laid out within the header bar.
+  final Widget? headerBarTrailing;
+
+  /// The list of navigation items.
+  final List<NavigationItem> items;
+
+  /// Called when one of the [items] is tapped.
+  final ValueChanged<int>? onItemSelected;
+
+  /// The index into [items] for the current active [MasterNavigationItem].
+  final int currentIndex;
 
   @override
   State<ArnaSideScaffold> createState() => _ArnaSideScaffoldState();
@@ -102,15 +118,21 @@ class _ArnaSideScaffoldState extends State<ArnaSideScaffold> {
                               onPressed: () =>
                                   _drawerOpenedCallback(!showDrawer),
                             ),
-                          widget.headerBarLeading,
-                          widget.items[_currentIndex].headerBarLeading,
+                          if (widget.headerBarLeading != null)
+                            widget.headerBarLeading!,
+                          if (widget.items[_currentIndex].headerBarLeading !=
+                              null)
+                            widget.items[_currentIndex].headerBarLeading!,
                         ],
                       ),
                       title: widget.title,
                       headerBarTrailing: Row(
                         children: [
-                          widget.items[_currentIndex].headerBarTrailing,
-                          widget.headerBarTrailing,
+                          if (widget.items[_currentIndex].headerBarTrailing !=
+                              null)
+                            widget.items[_currentIndex].headerBarTrailing!,
+                          if (widget.headerBarTrailing != null)
+                            widget.headerBarTrailing!,
                         ],
                       ),
                       searchField: widget.items[_currentIndex].searchField,
@@ -163,15 +185,17 @@ class _ArnaSideScaffoldState extends State<ArnaSideScaffold> {
         return ArnaScaffold(
           headerBarLeading: Row(
             children: [
-              widget.headerBarLeading,
-              widget.items[_currentIndex].headerBarLeading,
+              if (widget.headerBarLeading != null) widget.headerBarLeading!,
+              if (widget.items[_currentIndex].headerBarLeading != null)
+                widget.items[_currentIndex].headerBarLeading!,
             ],
           ),
           title: widget.title,
           headerBarTrailing: Row(
             children: [
-              widget.items[_currentIndex].headerBarTrailing,
-              widget.headerBarTrailing,
+              if (widget.items[_currentIndex].headerBarTrailing != null)
+                widget.items[_currentIndex].headerBarTrailing!,
+              if (widget.headerBarTrailing != null) widget.headerBarTrailing!,
             ],
           ),
           searchField: widget.items[_currentIndex].searchField,
@@ -183,25 +207,13 @@ class _ArnaSideScaffoldState extends State<ArnaSideScaffold> {
 }
 
 class NavigationItem {
-  final String title;
-  final IconData icon;
-  final WidgetBuilder builder;
-  final Widget headerBarLeading;
-  final Widget headerBarTrailing;
-  final ArnaSearchField? searchField;
-  final ArnaBadge? badge;
-  final bool isFocusable;
-  final bool autofocus;
-  final Color accentColor;
-  final MouseCursor cursor;
-  final String? semanticLabel;
-
+  /// Creates a navigation item.
   const NavigationItem({
     required this.title,
     required this.icon,
     required this.builder,
-    this.headerBarLeading = const SizedBox.shrink(),
-    this.headerBarTrailing = const SizedBox.shrink(),
+    this.headerBarLeading,
+    this.headerBarTrailing,
     this.searchField,
     this.badge,
     this.isFocusable = true,
@@ -210,4 +222,42 @@ class NavigationItem {
     this.cursor = MouseCursor.defer,
     this.semanticLabel,
   });
+
+  /// The title of the item.
+  final String title;
+
+  /// The icon of the item.
+  final IconData icon;
+
+  /// The widget builder of the item.
+  final WidgetBuilder builder;
+
+  /// The leading widget laid out within the header bar.
+  final Widget? headerBarLeading;
+
+  /// The trailing widget laid out within the header bar.
+  final Widget? headerBarTrailing;
+
+  /// The [ArnaSearchField] of the item.
+  final ArnaSearchField? searchField;
+
+  /// The [ArnaBadge] of the item.
+  final ArnaBadge? badge;
+
+  /// Whether this item is focusable or not.
+  final bool isFocusable;
+
+  /// Whether this item should focus itself if nothing else is already
+  /// focused.
+  final bool autofocus;
+
+  /// The color of the item's focused border.
+  final Color accentColor;
+
+  /// The cursor for a mouse pointer when it enters or is hovering over the
+  /// item.
+  final MouseCursor cursor;
+
+  /// The semantic label of the item.
+  final String? semanticLabel;
 }