+++ lib/src/dialogs/dialog.dart
@@ -110,7 +110,7 @@ Future<T?> showArnaDialog<T>({
   required BuildContext context,
   required WidgetBuilder builder,
   bool barrierDismissible = false,
-  Color barrierColor = Styles.barrierColor,
+  Color barrierColor = ArnaColors.barrierColor,
   String? barrierLabel = "label",
   bool useRootNavigator = true,
   RouteSettings? routeSettings,